GUI-1185: Prevent missing root device lookup from breaking image deregistration

+++ eucaconsole/views/images.py
@@ -286,13 +286,14 @@ class ImageView(TaggedItemView):
         if self.deregister_form.validate():
             with boto_error_handler(self.request):
                 delete_snapshot = False
+                root_dev = None
                 if self.image.root_device_type == 'ebs' and self.request.params.get('delete_snapshot') == 'y':
                     delete_snapshot = True
                     root_dev = panels.get_root_device_name(self.image)
                 self.conn.deregister_image(self.image.id)
                 if delete_snapshot:
                     for key in self.image.block_device_mapping:
-                        if key == root_dev:
+                        if root_dev and key == root_dev:
                             snapshot_id = self.image.block_device_mapping[key].snapshot_id
                             self.conn.delete_snapshot(snapshot_id)
                             break
